# api/src/swagger.yml
openapi: 3.0.0
info:
  title: Fake Booking API
  description: API for a take-home project
  version: 1.0.0

paths:
  /booking/shop/{shop}:
    get:
      summary: Get shop details
      parameters:
        - name: shop
          in: path
          required: true
          description: The unique identifier for the shop
          schema:
            type: string
            example: "shop123"
      responses:
        "200":
          description: A list of shops
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopResponse"

  /booking/user:
    get:
      summary: Get user details
      security:
        - BearerAuth: []
      responses:
        "200":
          description: A list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

  /booking/menu:
    post:
      summary: Reserve a menu item
      description: Reserve a specific menu item with quantity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuReservationRequest"
      responses:
        "200":
          description: Reservation confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuReservationResponse"
        "400":
          description: Invalid input
        "500":
          description: Server error

  /booking/reservation/{code}:
    get:
      summary: Get reservation details by ID
      parameters:
        - name: code
          in: path
          required: true
          description: The unique identifier for the reservation
          schema:
            type: string
            example: "res456"
      responses:
        "200":
          description: A list of reservations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"

  /booking/waitlist:
    post:
      summary: Join the waitlist
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinWaitlistInput"
      responses:
        "200":
          description: Waitlist response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinWaitlistResponse"

  /booking/waitlist/{id}:
    get:
      summary: Get the waitlist status by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier for the waitlist entry
          schema:
            type: string
            example: "wait789"
      responses:
        "200":
          description: Waitlist status response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinWaitlistResponse"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    MenuReservationRequest:
      type: object
      required:
        - id
        - quantity
      properties:
        id:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        quantity:
          type: integer
          minimum: 1
          example: 2
        notes:
          type: string
          description: Optional notes from user
          example: "No onions, please"

    MenuReservationResponse:
      type: object
      properties:
        message:
          type: string
          example: "Reservation received"
        reservationId:
          type: string
          example: "H8OIE0"

    MenuItem:
      type: object
      properties:
        id:
          type: string
          example: "menu001"
        image:
          type: string
          example: "https://thecoffeehouse.com/image1.jpg"
        name:
          type: string
          example: "Cheeseburger"
        description:
          type: string
          example: "A delicious cheeseburger with all the fixings."
        minQty:
          type: integer
          example: 1
        maxQty:
          type: integer
          example: 5
        price:
          type: string
          example: "9.99"
        currency:
          type: string
          example: "USD"
    ShopResponse:
      type: object
      properties:
        id:
          type: string
          example: "shop123"
        name:
          type: string
          example: "The Coffee House"
        bookingMinPax:
          type: integer
          example: 1
        bookingMaxPax:
          type: integer
          example: 10
        bookingTitle:
          type: string
          example: "Reserve at The Coffee House"
        detailsTitle:
          type: string
          example: "Your reservation details at The Coffee House"
        waitlistTitle:
          type: string
          example: "Join the waitlist at The Coffee House"
        menuTitle:
          type: string
          example: "Menu at The Coffee House"
        location:
          type: string
          example: "123 Main St, Anytown, USA"
        rating:
          type: number
          example: 4.5
        categories:
          type: array
          items:
            type: string
          example: ["Cafe", "Bakery"]
        hours:
          type: string
          example: "8 AM - 8 PM"
        contact:
          type: string
          example: "(555) 123-4567"
        website:
          type: string
          example: "https://thecoffeehouse.com"
        description:
          type: string
          example: "A cozy place to enjoy coffee and pastries."
        images:
          type: array
          items:
            type: string
          example:
            [
              "https://thecoffeehouse.com/image1.jpg",
              "https://thecoffeehouse.com/image2.jpg",
            ]
        menu:
          type: array
          items:
            $ref: "#/components/schemas/MenuItem"
          example: [{ id: "menu001", name: "Cheeseburger", price: "9.99" }]
        services:
          type: array
          items:
            type: string
          example: ["WiFi", "Outdoor Seating"]
        bookingTheme:
          type: object
          $ref: "#/components/schemas/BookingTheme"
        reviews:
          type: array
          items:
            type: string
          example: ["Great coffee!", "Loved the ambiance."]
        owner:
          type: string
          example: "John Doe"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-02T10:00:00Z"

    UserResponse:
      type: object
      properties:
        id:
          type: string
          example: "user123"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          example: "johndoe@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        phone:
          type: string
          example: "(555) 987-6543"
        address:
          type: string
          example: "456 Elm St, Anytown, USA"
        preferences:
          type: array
          items:
            type: string
          example: ["Vegan", "Gluten-Free"]
        loyaltyPoints:
          type: integer
          example: 1500
        membershipStatus:
          type: string
          example: "Gold"
        bookings:
          type: array
          items:
            type: string
          example: ["res123", "res456"]
        reviews:
          type: array
          items:
            type: string
          example: ["Great service!", "Will visit again."]
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-02T10:00:00Z"
        lastLogin:
          type: string
          format: date-time
          example: "2023-01-03T10:00:00Z"

    ReservationResponse:
      type: object
      properties:
        id:
          type: string
          example: "res123"
        shopId:
          type: string
          example: "shop123"
        date:
          type: string
          format: date
          example: "2023-01-10"
        time:
          type: string
          example: "12:00 PM"
        status:
          type: string
          example: "Confirmed"
        guests:
          type: integer
          example: 2
        specialRequests:
          type: string
          example: "Window seat"
        code:
          type: string
          example: "CONF123456"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-02T10:00:00Z"
        cancellationPolicy:
          type: string
          example: "24-hour notice required"
        paymentStatus:
          type: string
          example: "Paid"
        totalCost:
          type: number
          example: 49.99
        discounts:
          type: array
          items:
            type: string
          example: ["10% off"]

    JoinWaitlistInput:
      type: object
      properties:
        userId:
          type: string
          example: "user123"
        shopId:
          type: string
          example: "shop123"
        desiredDate:
          type: string
          format: date
          example: "2023-01-15"
        desiredTime:
          type: string
          example: "6:00 PM"
        partySize:
          type: integer
          example: 4
        contactPreference:
          type: string
          example: "Email"
        specialRequests:
          type: string
          example: "High chair needed"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T10:00:00Z"

    JoinWaitlistResponse:
      type: object
      properties:
        waitlistId:
          type: string
          example: "wait123"
        userId:
          type: string
          example: "user123"
        shopId:
          type: string
          example: "shop123"
        position:
          type: integer
          example: 5
        estimatedWaitTime:
          type: string
          example: "30 minutes"
        status:
          type: string
          example: "Waiting"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-02T10:00:00Z"
        notificationSent:
          type: boolean
          example: true
        code:
          type: string
          example: "CONF789012"
        specialRequests:
          type: string
          example: "Near entrance"
        contactPreference:
          type: string
          example: "SMS"
        partySize:
          type: integer
          example: 4
        desiredDate:
          type: string
          format: date
          example: "2023-01-15"
        desiredTime:
          type: string
          example: "6:00 PM"

    BookingTheme:
      type: object
      properties:
        colorPrimary:
          type: string
          format: color
          example: "#FF5733"
        colorSecondary:
          type: string
          format: color
          example: "#C70039"
        backgroundColor:
          type: string
          format: color
          example: "#FFFFFF"
        textColor:
          type: string
          format: color
          example: "#333333"
        borderRadius:
          type: string
          example: "8px"
        fontSize:
          type: string
          example: "1rem"
        spacing:
          type: string
          example: "1.5rem"